Fix some compiler warnings
